+++ src/main/java/dev/abarmin/bots/entity/telegram/TelegramBotUpdate.java
@@ -1,4 +1,4 @@
-package dev.abarmin.bots.entity;
+package dev.abarmin.bots.entity.telegram;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
@@ -7,16 +7,13 @@ import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
-import java.net.URI;
 import java.time.LocalDateTime;
 
-@Table("ARTICLES")
-public record Article(
-        @Column("ARTICLE_ID") @Id Integer id,
-        @Column("ARTICLE_SOURCE_ID") AggregateReference<ArticleSource, Integer> articleSource,
-        @Column("ARTICLE_TITLE") String articleTitle,
-        @Column("ARTICLE_URL") URI articleUri,
-        @Column("ARTICLE_ADDED") LocalDateTime articleAdded,
+@Table("TELEGRAM_UPDATES")
+public record TelegramBotUpdate(
+        @Column("ID") @Id Integer id,
+        @Column("UPDATE_ID") Integer updateId,
+        @Column("CHAT_ID") AggregateReference<TelegramBotChat, Long> chatId,
         @Column("CREATED_AT") @CreatedDate LocalDateTime createdAt,
         @Column("UPDATED_AT") @LastModifiedDate LocalDateTime updatedAt
         ) {
