+++ src/main/java/dev/abarmin/bots/repository/ArticleSubscriptionRepository.java
@@ -1,8 +1,8 @@
 package dev.abarmin.bots.repository;
 
-import dev.abarmin.bots.entity.ArticleSubscription;
-import dev.abarmin.bots.entity.TelegramBotChat;
-import dev.abarmin.bots.entity.ArticleSource;
+import dev.abarmin.bots.entity.rss.ArticleSubscription;
+import dev.abarmin.bots.entity.telegram.TelegramBotChat;
+import dev.abarmin.bots.entity.rss.ArticleSource;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.repository.ListCrudRepository;
