+++ src/main/java/dev/abarmin/bots/repository/TelegramBotChatRepository.java
@@ -1,6 +1,6 @@
 package dev.abarmin.bots.repository;
 
-import dev.abarmin.bots.entity.TelegramBotChat;
+import dev.abarmin.bots.entity.telegram.TelegramBotChat;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.data.repository.ListPagingAndSortingRepository;
