+++ src/main/java/dev/abarmin/bots/service/RecommendationService.java
@@ -1,6 +1,6 @@
 package dev.abarmin.bots.service;
 
-import dev.abarmin.bots.entity.ArticleSource;
+import dev.abarmin.bots.entity.rss.ArticleSource;
 
 import java.util.Collection;
