+++ src/main/java/dev/abarmin/bots/service/SubscriptionService.java
@@ -1,8 +1,8 @@
 package dev.abarmin.bots.service;
 
-import dev.abarmin.bots.entity.TelegramBotChat;
-import dev.abarmin.bots.entity.ArticleSubscription;
-import dev.abarmin.bots.entity.ArticleSource;
+import dev.abarmin.bots.entity.telegram.TelegramBotChat;
+import dev.abarmin.bots.entity.rss.ArticleSubscription;
+import dev.abarmin.bots.entity.rss.ArticleSource;
 
 import java.net.URI;
 import java.util.Collection;
