+++ src/main/java/dev/abarmin/bots/service/digest/DeleteSubscriptionStateHandler.java
@@ -7,11 +7,11 @@ import com.pengrad.telegrambot.request.SendMessage;
 import dev.abarmin.bots.service.support.BotHelper;
 import dev.abarmin.bots.service.support.BotOperation;
 import dev.abarmin.bots.service.support.MessageSourceHelper;
-import dev.abarmin.bots.entity.TelegramBotChat;
+import dev.abarmin.bots.entity.telegram.TelegramBotChat;
 import dev.abarmin.bots.service.TelegramChatService;
 import dev.abarmin.bots.model.DigestBotUpdate;
 import dev.abarmin.bots.service.SubscriptionService;
-import dev.abarmin.bots.entity.ArticleSource;
+import dev.abarmin.bots.entity.rss.ArticleSource;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.context.ApplicationEventPublisher;
