+++ src/main/java/dev/abarmin/bots/service/impl/SubscriptionServiceImpl.java
@@ -1,9 +1,9 @@
 package dev.abarmin.bots.service.impl;
 
-import dev.abarmin.bots.entity.TelegramBotChat;
-import dev.abarmin.bots.entity.ArticleSubscription;
+import dev.abarmin.bots.entity.telegram.TelegramBotChat;
+import dev.abarmin.bots.entity.rss.ArticleSubscription;
 import dev.abarmin.bots.repository.ArticleSubscriptionRepository;
-import dev.abarmin.bots.entity.ArticleSource;
+import dev.abarmin.bots.entity.rss.ArticleSource;
 import dev.abarmin.bots.repository.ArticleSourceRepository;
 import dev.abarmin.bots.service.ArticleSourceService;
 import dev.abarmin.bots.service.SubscriptionService;
